LASIR-T1-D1-c -- Implements LASIR Task #1, Deliverable 1c: Requested Google Scholar SEO enhancements
Requirements:

The purpose of this is to ensure that automatic crawlers are able to discover and fetch the URLs of all articles, as well as to periodically refresh their content. In Google Scholar setup (https://partnerdash.google.com/partnerdash/d/scholarinclusions#p:id=new), you are asked to provide the following:

Please provide the URL of a page listing all years available.
Please provide examples of URLs to pages that provide a list of articles for a given year.
Please provide examples of individual abstract pages.
Please provide examples of full-text article URLs. Please pick some recently added articles as well as some articles that were added a while ago.
The first two will require creating custom displays, while the third and fourth require no additional functionality. In addition, for larger repositories, GS recommends providing a page that lists items added in the last two weeks. So this part of the task requires building three pages:

A page that lists years. I think the way to do this is to have a custom page which runs a solr query returning zero rows, and faceted with a facet.date.gap of a year.
A page listing articles, filtered by a year parameter in the url.
A page showing articles added/updated in the past two weeks.

These requirements are intended to be met in this commit as follows:

- New pages added:
  - /gs_years: Lists all years when articles were published.
  - /gs_year/%: Lists all articles published in a given year
  - /gs_updated: Lists all articles added or updated in the repository in the past two weeks
  - /gs_abstract/%: Displays a very basic page showing the title and abstract for a given object
- Expanded islandora_scholar configuration form:
  - Added fields to the Google Scholar section for which solr fields to use for publication date and abstract, and which content models to exclude from these pages.
- Documented these enhancements in the islandora_google_scholar README file.